feat: unify astro pools
