+++ src/implementations/astroport/helpers.rs
@@ -1,4 +1,7 @@
-use astroport_core::{asset::{Asset as AstroAsset, AssetInfo as AstroAssetInfo}, U256};
+use astroport_core::{
+    asset::{Asset as AstroAsset, AssetInfo as AstroAssetInfo},
+    U256,
+};
 use cosmwasm_std::{StdError, StdResult};
 use cw_asset::{Asset, AssetInfo, AssetList};
 
@@ -80,4 +83,180 @@ pub fn checked_u8_mul(a: &U256, b: u8) -> Option<U256> {
         result = result.checked_add(*a)?;
     }
     Some(result)
+}
+
+use std::cmp::Ordering;
+
+use astroport_core::factory::PairType;
+use astroport_core::querier::{query_supply, query_token_precision};
+use cosmwasm_schema::cw_serde;
+use cosmwasm_std::{from_slice, Addr, Decimal, Env, QuerierWrapper, Response};
+use cosmwasm_std::{Deps, Uint128};
+use cw_asset::AssetInfoBase;
+use delegate::delegate;
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+
+use crate::traits::Pool;
+use crate::CwDexError;
+use astroport_core::asset::PairInfo;
+
+pub(crate) const N_COINS: u8 = 2;
+pub const AMP_PRECISION: u64 = 100;
+const ITERATIONS: u8 = 32;
+
+/// ## Description
+/// Compute the current pool amplification coefficient (AMP).
+/// ## Params
+/// * **config** is an object of type [`Config`].
+///
+/// * **env** is an object of type [`Env`].
+///
+/// This function is needed to calculate how many LP shares a user should get when providing liquidity but is
+/// not publicly exposed in the package. Copied from the astro implementation here:
+/// https://github.com/astroport-fi/astroport-core/blob/c216ecd4f350113316be44d06a95569f451ac681/contracts/pair_stable/src/contract.rs#L1492-L1515
+pub(crate) fn compute_current_amp(config: &Config, env: &Env) -> StdResult<u64> {
+    let block_time = env.block.time.seconds();
+
+    if block_time < config.next_amp_time {
+        let elapsed_time =
+            Uint128::from(block_time).checked_sub(Uint128::from(config.init_amp_time))?;
+        let time_range =
+            Uint128::from(config.next_amp_time).checked_sub(Uint128::from(config.init_amp_time))?;
+        let init_amp = Uint128::from(config.init_amp);
+        let next_amp = Uint128::from(config.next_amp);
+
+        if config.next_amp > config.init_amp {
+            let amp_range = next_amp - init_amp;
+            let res = init_amp + (amp_range * elapsed_time).checked_div(time_range)?;
+            Ok(res.u128() as u64)
+        } else {
+            let amp_range = init_amp - next_amp;
+            let res = init_amp - (amp_range * elapsed_time).checked_div(time_range)?;
+            Ok(res.u128() as u64)
+        }
+    } else {
+        Ok(config.next_amp)
+    }
+}
+
+/// ## Description
+/// Return a value using a newly specified precision.
+/// ## Params
+/// * **value** is an object of type [`Uint128`]. This is the value that will have its precision adjusted.
+///
+/// * **current_precision** is an object of type [`u8`]. This is the `value`'s current precision
+///
+/// * **new_precision** is an object of type [`u8`]. This is the new precision to use when returning the `value`.
+///
+/// Copied from the astro code here:
+/// https://github.com/astroport-fi/astroport-core/blob/c216ecd4f350113316be44d06a95569f451ac681/contracts/pair_stable/src/contract.rs#L1269
+pub(crate) fn adjust_precision(
+    value: Uint128,
+    current_precision: u8,
+    new_precision: u8,
+) -> StdResult<Uint128> {
+    Ok(match current_precision.cmp(&new_precision) {
+        Ordering::Equal => value,
+        Ordering::Less => value
+            .checked_mul(Uint128::new(10_u128.pow((new_precision - current_precision) as u32)))?,
+        Ordering::Greater => value
+            .checked_div(Uint128::new(10_u128.pow((current_precision - new_precision) as u32)))?,
+    })
+}
+
+/// ## Description
+/// Computes stable swap invariant (D)
+///
+/// * **Equation**
+///
+/// A * sum(x_i) * n**n + D = A * D * n**n + D**(n+1) / (n**n * prod(x_i))
+///
+/// ## Params
+/// * **leverage** is the object of type [`u128`].
+///
+/// * **amount_a** is the object of type [`u128`].
+///
+/// * **amount_b** is the object of type [`u128`].
+pub(crate) fn compute_d(leverage: u64, amount_a: u128, amount_b: u128) -> Option<u128> {
+    let amount_a_times_coins =
+        checked_u8_mul(&U256::from(amount_a), N_COINS)?.checked_add(U256::one())?;
+    let amount_b_times_coins =
+        checked_u8_mul(&U256::from(amount_b), N_COINS)?.checked_add(U256::one())?;
+    let sum_x = amount_a.checked_add(amount_b)?; // sum(x_i), a.k.a S
+    if sum_x == 0 {
+        Some(0)
+    } else {
+        let mut d_previous: U256;
+        let mut d: U256 = sum_x.into();
+
+        // Newton's method to approximate D
+        for _ in 0..ITERATIONS {
+            let mut d_product = d;
+            d_product = d_product.checked_mul(d)?.checked_div(amount_a_times_coins)?;
+            d_product = d_product.checked_mul(d)?.checked_div(amount_b_times_coins)?;
+            d_previous = d;
+            //d = (leverage * sum_x + d_p * n_coins) * d / ((leverage - 1) * d + (n_coins + 1) * d_p);
+            d = calculate_step(&d, leverage, sum_x, &d_product)?;
+            // Equality with the precision of 1
+            if d == d_previous {
+                break;
+            }
+        }
+        u128::try_from(d).ok()
+    }
+}
+
+/// ## Description
+/// Calculates step
+///
+/// * **Equation**:
+///
+/// d = (leverage * sum_x + d_product * n_coins) * initial_d / ((leverage - 1) * initial_d + (n_coins + 1) * d_product)
+fn calculate_step(initial_d: &U256, leverage: u64, sum_x: u128, d_product: &U256) -> Option<U256> {
+    let leverage_mul = U256::from(leverage).checked_mul(sum_x.into())? / AMP_PRECISION;
+    let d_p_mul = checked_u8_mul(d_product, N_COINS)?;
+
+    let l_val = leverage_mul.checked_add(d_p_mul)?.checked_mul(*initial_d)?;
+
+    let leverage_sub =
+        initial_d.checked_mul((leverage.checked_sub(AMP_PRECISION)?).into())? / AMP_PRECISION;
+    let n_coins_sum = checked_u8_mul(d_product, N_COINS.checked_add(1)?)?;
+
+    let r_val = leverage_sub.checked_add(n_coins_sum)?;
+
+    l_val.checked_div(r_val)
+}
+
+// Astroport StableSwap pair does not return needed Config elements with smart query
+// Raw query gets all the necessary elements
+pub fn query_pair_config(querier: &QuerierWrapper, pair: &Addr) -> StdResult<Config> {
+    if let Some(res) = querier.query_wasm_raw(pair, b"config".as_slice())? {
+        let res: Config = from_slice(&res)?;
+        Ok(res)
+    } else {
+        Err(cosmwasm_std::StdError::GenericErr {
+            msg: "Raw query failed: config not found on pair address".to_string(),
+        })
+    }
+}
+
+/// ## Description
+/// This structure describes the main control config of pair stable.
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+pub struct Config {
+    /// the type of pair info available in [`PairInfo`]
+    pub pair_info: PairInfo,
+    /// the factory contract address
+    pub factory_addr: Addr,
+    /// The last time block
+    pub block_time_last: u64,
+    /// The last cumulative price 0 asset in pool
+    pub price0_cumulative_last: Uint128,
+    /// The last cumulative price 1 asset in pool
+    pub price1_cumulative_last: Uint128,
+    pub init_amp: u64,
+    pub init_amp_time: u64,
+    pub next_amp: u64,
+    pub next_amp_time: u64,
 }
