+++ src/implementations/astroport/mod.rs
@@ -1,9 +1,5 @@
-mod base_pool;
 mod helpers;
-mod stableswap;
+mod pool;
 mod staking;
-mod xyk;
 
-pub use stableswap::*;
 pub use staking::AstroportStaking;
-pub use xyk::AstroportXykPool;
